Move HashRouter to libxrpl
