+++ src/libxrpl/core/HashRouter.cpp
@@ -1,5 +1,4 @@
-#include <xrpld/app/misc/HashRouter.h>
-#include <xrpld/core/Config.h>
+#include <xrpl/core/HashRouter.h>
 
 namespace xrpl {
