+++ src/xrpld/app/tx/detail/Escrow.cpp
@@ -1,4 +1,3 @@
-#include <xrpld/app/misc/HashRouter.h>
 #include <xrpld/app/tx/detail/Escrow.h>
 #include <xrpld/app/tx/detail/MPTokenAuthorize.h>
 #include <xrpld/conditions/Condition.h>
@@ -6,6 +5,7 @@
 
 #include <xrpl/basics/Log.h>
 #include <xrpl/basics/chrono.h>
+#include <xrpl/core/HashRouter.h>
 #include <xrpl/ledger/ApplyView.h>
 #include <xrpl/ledger/CredentialHelpers.h>
 #include <xrpl/ledger/View.h>
